[Fullstack] EDITOR-#104 profile page
