+++ frontend/src/App.tsx
@@ -7,6 +7,7 @@ import Footer from "./components/Common/Footer.tsx";
 import useArticles from "./context/articleContext.tsx";
 import { UserContext } from "./context/userContext.tsx";
 import { useContext } from "react";
+import EditProfilePage from "./containers/User/EditProfilePage.tsx";
 
 const LoginPage = lazy(() => import("./containers/User/LoginPage.tsx"));
 const LogOutPage = lazy(() => import("./containers/User/LogOutPage.tsx"));
@@ -111,6 +112,14 @@ function App() {
             element={
               <Suspense fallback={<div>Loading...</div>}>
                 <ProfilePage />
+              </Suspense>
+            }
+          />
+          <Route
+            path="/user/:id/edit"
+            element={
+              <Suspense fallback={<div>Loading...</div>}>
+                <EditProfilePage />
               </Suspense>
             }
           />
