+++ frontend/src/containers/User/ProfilePage.tsx
@@ -1,58 +1,19 @@
-import { useLazyQuery, useMutation } from "@apollo/client";
+import { useLazyQuery, useMutation, useQuery } from "@apollo/client";
 import { useState } from "react";
-import {
-  ALL_USERS_QUERY,
-  UPDATE_USER_MUTATION,
-  UPDATE_USER_PROFILE_MUTATION,
-} from "../../graphql";
-import { useParams } from "react-router-dom";
+import { ALL_USERS_QUERY } from "../../graphql";
+import { Link, useParams } from "react-router-dom";
+import { User } from "@shared/shared_types";
 
 const ProfilePage = () => {
   const { id } = useParams();
-  if (!id) throw new Error("id is undefined");
-  const [edit, setEdit] = useState(false);
-  const [name, setName] = useState("");
-  const [intro, setIntor] = useState("");
-  const [studentID, setStudentID] = useState("");
-  const [photoLink, setPhotoLink] = useState("");
 
-  const [loadExpenseStatus, { loading: allLoading, error: allError }] =
-    useLazyQuery(ALL_USERS_QUERY);
-
-  const [updateUser, { loading, error }] = useMutation(
-    UPDATE_USER_PROFILE_MUTATION,
-  );
-
-  const handleUpdate = async () => {
-    if (!id) return;
-    if (!name) return;
-    if (!intro) return;
-    if (!studentID) return;
-    if (!photoLink) return;
-    if (loading) return "Submitting...";
-    if (error) return `Submission error! ${error.message}`;
-    console.log(name, intro);
-
-    const editedUser = await updateUser({
-      variables: {
-        updateUserProfileId: parseInt(id),
-        userProfileInput: {
-          name: name,
-          introduction: intro,
-          photoLink: photoLink,
-          studentID: studentID,
-        },
-      },
-    });
-    setEdit(false);
-    if (allLoading) return <div>Loading...</div>;
-    if (allError) return <div>{allError.message}</div>;
-    let queryUser = await loadExpenseStatus();
-    console.log(queryUser.data?.AllUsers);
-    console.log(queryUser.data?.AllUsers);
-    window.location.reload();
-  };
-
+  const { loading, error, data } = useQuery(ALL_USERS_QUERY);
+  if (loading) return;
+  if (error) return `Error! ${error?.message}`;
+  const users = JSON.parse(JSON.stringify(data?.AllUsers));
+  if (!id) return;
+  const user = users.filter((user: User) => `${user.id}` === id)[0];
+  console.log(user);
   return (
     <div className="flex rounded bg-white mt-5 mb-5 justify-center">
       <div className="flex flex-row w-fit">
@@ -61,100 +22,47 @@ const ProfilePage = () => {
             <img
               className="rounded-circle mt-5"
               width="150px"
-              src="https://st3.depositphotos.com/15648834/17930/v/600/depositphotos_179308454-stock-illustration-unknown-person-silhouette-glasses-profile.jpg"
+              src={user.photoLink}
+              alt="not found"
             />
           </div>
         </div>
         <div className="col-md-5 border-right">
           <div className="p-3 py-5">
-            {!edit && (
-              <button
-                className="text-md font-bold bg-blue-700 mt-0 text-white px-4 py-1 rounded hover:bg-blue-700 mr-2 "
-                onClick={() => setEdit(true)}
-              >
+            <Link to={`./edit`}>
+              <button className="text-md font-bold bg-blue-700 mt-0 text-white px-4 py-1 rounded hover:bg-blue-700 mr-2 ">
                 edit
               </button>
-            )}
-            {edit && (
-              <button
-                className="text-md font-bold bg-blue-700 mt-0 text-white px-4 py-1 rounded hover:bg-blue-700 mr-2 "
-                onClick={() => handleUpdate()}
-              >
-                save
-              </button>
-            )}
+            </Link>
             <div className="flex mt-2">
               <span className="bg-slate-400 p-1 px-4 rounded text-white center">
                 Name
               </span>
             </div>
-            {!edit && <p className="p-1 px-4 rounded">name</p>}
-            {edit && (
-              <input
-                className="p-1 px-4 rounded"
-                type="text"
-                id="name"
-                value={name}
-                onChange={(event) => {
-                  setName(event.target.value);
-                }}
-                placeholder="Search something.."
-              />
-            )}
+            <p className="p-1 px-4 rounded">{user.name}</p>
 
             <div className="flex flex-row space-x-2">
               <span className="bg-slate-400 p-1 px-4 rounded text-white center">
                 Introduction
               </span>
             </div>
-            {!edit && <p className="p-1 px-4 rounded">{":)"}</p>}
-            {edit && (
-              <textarea
-                value={intro}
-                onChange={(event) => {
-                  setIntor(event.target.value);
-                }}
-              ></textarea>
-            )}
+            <p className="p-1 px-4 rounded">
+              {user.introduction ? user.introduction : ":)"}
+            </p>
+
             <div className="flex mt-2">
               <span className="bg-slate-400 p-1 px-4 rounded text-white center">
                 StudentId
               </span>
             </div>
-            {!edit && <p className="p-1 px-4 rounded">bxxxxxxxx</p>}
-            {edit && (
-              <input
-                className="p-1 px-4 rounded"
-                type="text"
-                id="photoLink"
-                value={studentID}
-                onChange={(event) => {
-                  setStudentID(event.target.value);
-                }}
-                placeholder="Search something.."
-              />
-            )}
+            <p className="p-1 px-4 rounded">{user.studentID}</p>
 
-            <div className="flex mt-2">
+            {/* <div className="flex mt-2">
               <span className="bg-slate-400 p-1 px-4 rounded text-white center">
                 PhotoLink
               </span>
             </div>
-            {!edit && (
-              <p className="p-1 px-4 rounded">https://www.google.com/</p>
-            )}
-            {edit && (
-              <input
-                className="p-1 px-4 rounded"
-                type="text"
-                id="photoLink"
-                value={photoLink}
-                onChange={(event) => {
-                  setPhotoLink(event.target.value);
-                }}
-                placeholder="Search something.."
-              />
-            )}
+            <p className="p-1 px-4 rounded">{user.photoLink}</p> */}
           </div>
         </div>
       </div>
